% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damb.R
\docType{data}
\name{damb}
\alias{damb}
\title{example of DATRAS HH file from SP-PORC survey in 2016}
\format{a data.frame with 83 observations (hauls) of 61 variables 
      See DATRAS docs on HH files}
\usage{
data(damb)
}
\description{
HH file with data from all the haul info from the Spanish Porcupine Survey 
in 2016, This data frame is used in examples to facilitate them when there's 
no Internet conection
}
\keyword{datasets}
