% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHLdata.ts.R
\name{getHLdata.ts}
\alias{getHLdata.ts}
\title{Function getHLdata.ts to download from DATRAS HL records getting several years in a row, uses getHHdata from icesDatras that only gets one year}
\usage{
getHLdata.ts(Survey, years, quarter)
}
\arguments{
\item{Survey:}{Surveys available in DATRAS see details}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in yellow, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be ploted}
}
\description{
Function getHLdata.ts to download from DATRAS HL records getting several years in a row, uses getHHdata from icesDatras that only gets one year
}
\details{
Surveys available in DATRAS: i.e. SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
getHLdata.ts("SP-NORTH",c(2014:2016),4)
}
