% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBTSNeAtl_map.R
\name{IBTSNeAtl_map}
\alias{IBTSNeAtl_map}
\title{Function IBTSNeAtl_map plots the map with all the surveys **STILL ON DEVELOPMENT**}
\usage{
IBTSNeAtl_map(
  nl = 60.5,
  sl = 36,
  xlims = c(-18, 3),
  leg = TRUE,
  cex.leg = 0.7,
  dens = 30,
  load = TRUE,
  ICESdiv = TRUE,
  ICESrect = FALSE,
  NS = TRUE,
  bathy = TRUE,
  bw = FALSE,
  axlab = 0.8,
  bords = TRUE,
  out = "def",
  nfile = "NeAtlIBTS_map",
  lwdl = 0.1,
  shpdir = "c:/GitHubRs/shapes/"
)
}
\arguments{
\item{nl}{= 61.5 northernmost limit of the map}

\item{sl}{= 35 Southernmost limit of the map}

\item{leg}{= TRUE if TRUE includes the legend with the colors of the surveys}

\item{cex.leg}{= .7 Size of the legend}

\item{dens}{= 30 density of the shading lines for all the surveys}

\item{load}{= T or F to load all the shapes files.}

\item{ICESdiv}{= TRUE if TRUE plots the IBTS divisions behind the shapefiles}

\item{ICESrect}{= FALSE if TRUE plots the lines of the ICES statistic rectangles}

\item{NS}{= FALSE if TRUE includes the ICES rectangles only for the North Sea area}

\item{bathy}{= TRUE if TRUE plots the isobaths under the behind the shapefiles}

\item{bw}{= True plots the map with land in grey, if F in light brown (burlywood3)}

\item{bords}{= TRUE plots the borders o the countries, FALSE leaves dashed lines}

\item{out}{= format of the output, can be "def" default device, "pdf", "tiff" or "png"}

\item{nfile}{= name for the output file}

\item{shpdir}{= path to the folder with the shapefiles}

\item{axlab=}{.8 decides the size of the axis numbers}
}
\description{
Produces a map from the shapefiles that define the IBTSNeAtl Surveys from Scotland to Cádiz, Still needs to include the shapefiles within the package, or right references.
}
\examples{
IBTSNeAtl_map(out="def",dens=0,nl=45,leg=F,load=TRUE,ICESrect = T);text(stat_y~stat_x,Area,labels=ICESNAME,cex=.8,font=4);text(stat_y~stat_x,Area,labels=Area,cex=.6,pos=1,col=2) 
}
