% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesCAperYear.R
\name{SpeciesCAperYear}
\alias{SpeciesCAperYear}
\title{Function SpeciesCAperYear Takes a CA file for one survey and produces a table with number of biological samples per year and species}
\usage{
SpeciesCAperYear(CA)
}
\arguments{
\item{CA:}{A data frame with data from CA samples got with getCAdata.ts}
}
\description{
Function SpeciesCAperYear Takes a CA file for one survey and produces a table with number of biological samples per year and species
}
\examples{
SpeciesCAperYear(getCAdata.ts("SP-NORTH",c(2011:2012),4))
}
