% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyMap.IBTS.R
\name{SurveyMap.IBTS}
\alias{SurveyMap.IBTS}
\title{Map of hauls (stations) from an IBTS survey taken directly from DATRAS}
\usage{
SurveyMap.IBTS(
  Survey,
  Year,
  Quarter,
  ti = TRUE,
  leg = TRUE,
  legcount = FALSE,
  ICESrect = FALSE,
  ICESlab = FALSE,
  ICESlabcex = 0.7
)
}
\arguments{
\item{Survey}{Code to plot. Has to be one of those available from DATRAS survey list, and from a year and quarter avalable}

\item{Year}{year to plot}

\item{Quarter}{to plot}

\item{ti}{TRUE includes the Survey Year Quarter title by default, a text includes the txt as title}

\item{leg}{TRUE includes a legend with the country-colors codes}
}
\value{
Presents the map with the stations and a legend with the countries that have participated in the survey
}
\description{
Uses the DATRAS HH data of a survey, year and quarter to plot stations in a map
}
\examples{
SurveyMap.IBTS("NS-IBTS",2021,3)
SurveyMap.ibts("FR-WCGFS",2303,3,ICESrect = T,ICESlab = T,ICESlabcex = .6)
}
\concept{maps}
