% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBTSNeAtl_map.R
\name{IBTSNeAtl_map}
\alias{IBTSNeAtl_map}
\title{Function IBTSNeAtl_map plots the map with all the surveys **STILL ON DEVELOPMENT**}
\usage{
IBTSNeAtl_map(nl = 60.5, sl = 36, xlims = c(-18, 3), leg = TRUE,
  cex.leg = 0.7, dens = 30, ICESdiv = TRUE, icesLab = TRUE,
  bathy = TRUE, out = "tiff", nfile = "NeAtlIBTS_map", lwdl = 0.1,
  shpdir = "c:/users/francisco.velasco.ST/documents/fvg/campañas/IBTS/shapes/")
}
\arguments{
\item{nl}{= 61.5 northernmost limit of the map}

\item{sl}{= 35 Southernmost limit of the map}

\item{leg}{= TRUE if TRUE includes the legend with the colors of the surveys}

\item{cex.leg}{= .7 Size of the legend}

\item{dens}{= 30 density of the shading lines for all the surveys}

\item{ICESdiv}{= TRUE if TRUE plots the IBTS divisions behind the shapefiles}

\item{bathy}{= TRUE if TRUE plots the isobaths under the behind the shapefiles}

\item{out}{= format of the output, can be "def" default device, "pdf", "tiff" or "png"}

\item{nfile}{= name for the output file}

\item{shpdir}{= path to the folder with the shapefiles}
}
\description{
Produces a map from the shapefiles that define the IBTSNeAtl Surveys from Scotland to Cádiz, Still needs to include the shapefiles within the package, or right references.
}
